improvement(build), block the build immediately when component-issues are found and these issues are tag blockers
